Release (#269)
* Change pre release back to develop and version

* support consumer counter in StreamHandlers

* Allow json content type on create policy route (POST)

* README updates

* README: add link to current metrics and sample output

* fix some container and build issues (#200)

* upgrade to latest gocui, fixes mouse bug when run from container

* fix pktvisor-cli docker build and runtime

* rename pktvisord-slim docker image to pktvisord

* add option to run only version code

* Clear filter before stopping pcapDevice

* log sflow biding

* improve json info in policies (#206)

* fix global configuration of period and deep sample for window config (#211)

* fix global configuration of period and deep sample for window config
* explicit template type to fix linux

* cli: run gofmt on version.go.in

* build: allow using existing conan build and host profile

* build: sort conanfile and remove generators

* build: workflow for static compilation

* Add retry in module child to get proper information when parsing json array

 - Yaml Cpp does not properly parse json array as expected by policy parser

* build: run cross compilation for develop branch

* Create dnstap_msg_type filter in DNS handler

* Implement filter by host mask for Dnstap Input

* Add support to metrics in policy and create GroupMetrics in Handler

* Add groups support to handlers and handler groups in DNS Handler

* add configure group method and not pass that in constructor

* use bitset operator[] to reduce overhead

* Validate filter by host and fix match ips

* refactor DNS Handler class to be possible to match groups in one place

* Refactor Net Handler to process input data in one method

* add groups on net handler

* Add bitset filter to dnstap

* refactor some group handling to base class

* Filter only if both response and query ip adresses don't match

* fix filtered dnstap packets

* Add unit tests for filtering dnstap

* Add doc messages on filter by hostmask

* move filtering logic to _filtering method

* Add support to metrics in policy and create GroupMetrics in Handler

* Add groups support to handlers and handler groups in DNS Handler

* add configure group method and not pass that in constructor

* use bitset operator[] to reduce overhead

* refactor DNS Handler class to be possible to match groups in one place

* Refactor Net Handler to process input data in one method

* add groups on net handler

* refactor some group handling to base class

* change metrics key to metric_groups and add method to check if group is enabled

* remove definition of MetricGroupIntType from streamHandler

* remove not necessary validation

* inline _filtering method in DNSTAP input stream

* Add unit tests for DNS handler groups

* add group unit tests for net handler

* validate group when merging metrics

* create unit tests for metric groups in policy class

* fix pktvisor-cli version warning (#220)


* replace bullseye-slim with bookworm-slim

* Revert "replace bullseye-slim with bookworm-slim"

* Add proper indentation for chain handlers

* Update pktvisor code to use latest version of pcapplusplus library

* break after getting pcap live device instance

* remove not necessary internal pcap operations

* Add new_event to pcap handler callbacks to allow new bucket creation (#227)

* Update 3rd party libs (#231)

* Update 3rd party libs

    - Update conan lib version to latest available
    - Update datasketches to version 3.3.0

* restore datasketches Cmake files and UVW lib version

* Add gdb to docker debug (#234)

* gdb should be available at runtime for debug image (#235)

* Modify pktvisor CMake to allow it to be added as subdirectory in other projects (#236)

* replace cmake_source_dir with project_source on integration tests (#239)

* starting directory for pktvisor automated tests (#237)

* starting repo for pktvisor automated tests

* boostrap tests

* Implement ThreadMonitor for linux systems (#230)

* Implement ThreadMonitor for linux systems

* Resources metrics as Handler

* Implement cpu percentage based on htop program

* reuse pcap timestamp to check timediff

* Add Resources Handler to every new input stream

* add unit tests for resources handler

* add coverage to prometheus method

* rename resources handler to input resources handler

* fix typo

* create new policy with resource handler for each new input

* remove not necessary include from CoreServer

* Add unique name for resources handler

* add input resources policy to input policy list

* only add input resource policy if the creation policy succeed

* verify if input resources policy exists before removing it

* start input resources policy only after the added policy

* do proper roll back policy

* Add support to count policies and handlers attached to input stream

* Properly handle policies and handlers count on input resources handler

* Only merge the fist bucket which is the more recent one

* Change variables name

* improve code coverage of input resources module

* Ensure that resources policy pointer is initialized as nullptr (#245)

* Tests/policies (#248)

get, create and remove a policy

* Feature/adjust qname aggregation (#242)

* Threat static sufix on DNS Qnames

* compare without casting

* add only_qname_suffix validation on unit tests

* replace string_view with size_t for suffix

* Fix aggregation tests

* Implement tcp timeout with LRU list (#249)

* Implement tcp timeout with LRU list

* put connection on LRUList on start callback

* local_thread cached data (#243)

* local_thread cached data

* replace pair with struct

* dnsLayer should exist

* perform cached data logic only in Handler class

* thread_local member class

* Add lower case qname method on DnsResource

* Add support to robin hood map to dns transactions

* policies scenarios description (#251)

* policies scenarios description

* Don't limit dns wire data by size (#254)

* Don't limit dns wire data by size

* Check buffer size minimum

* user permission and single handler (#257)

* Avoid undefined behavior on LRUList (#258)

* Add support to packet size and throughput (#261)

* Add support to packet size and throughput

* rename size varible to payload_size

* avoid send not tcp dns packet (#262)

* Feature/support to crashpad trace (#260)

* Add support to stacktrace with crashpad

* Handle errors on crashpad setup

* Add own crashpad conan code

* Exit with failure if missing crashpad information

* add support to disable crashpad and fix unit tests

* add back the conan server crashpad

* Add crash handler on debug images (#265)

* [FIX] adding step to build symbol
* [FIX] crashpad final path

* [FIX] docker entrypoint (#268)

* [FIX] docker entrypoint

* [FIX] removing garbage

* [FIX] sleep to suport crashpad submit

* bump release 4.1.0

+++ 3rd/datasketches/common/include/binomial_bounds.hpp
@@ -381,7 +381,7 @@ private:
   // The following computes an approximation to the lower bound of a Frequentist
   // confidence interval based on the tails of the Binomial distribution.
   static double compute_approx_binomial_lower_bound(unsigned long long num_samples, double theta, unsigned num_std_devs) {
-    if (theta == 1) return num_samples;
+    if (theta == 1) return static_cast<double>(num_samples);
     if (num_samples == 0) return 0;
     if (num_samples == 1) {
       const double delta = delta_of_num_std_devs[num_std_devs];
@@ -395,24 +395,24 @@ private:
     }
     // at this point we know 2 <= num_samples <= 120
     if (theta > (1 - 1e-5)) { // empirically-determined threshold
-      return num_samples;
+      return static_cast<double>(num_samples);
     }
     if (theta < (num_samples / 360.0)) { // empirically-determined threshold
       // here we use the Gaussian approximation, but with a modified num_std_devs
-      const unsigned index = 3 * num_samples + (num_std_devs - 1);
+      const unsigned index = 3 * static_cast<unsigned>(num_samples) + (num_std_devs - 1);
       const double raw_lb = cont_classic_lb(num_samples, theta, lb_equiv_table[index]);
       return raw_lb - 0.5; // fake round down
     }
     // This is the most difficult range to approximate; we will compute an "exact" LB.
     // We know that est <= 360, so specialNStar() shouldn't be ridiculously slow.
     const double delta = delta_of_num_std_devs[num_std_devs];
-    return special_n_star(num_samples, theta, delta); // no need to round
+    return static_cast<double>(special_n_star(num_samples, theta, delta)); // no need to round
   }
 
   // The following computes an approximation to the upper bound of a Frequentist
   // confidence interval based on the tails of the Binomial distribution.
   static double compute_approx_binomial_upper_bound(unsigned long long num_samples, double theta, unsigned num_std_devs) {
-    if (theta == 1) return num_samples;
+    if (theta == 1) return static_cast<double>(num_samples);
     if (num_samples == 0) {
       const double delta = delta_of_num_std_devs[num_std_devs];
       const double raw_ub = std::log(delta) / std::log(1 - theta);
@@ -425,18 +425,18 @@ private:
     }
     // at this point we know 2 <= num_samples <= 120
     if (theta > (1 - 1e-5)) { // empirically-determined threshold
-      return num_samples + 1;
+      return static_cast<double>(num_samples + 1);
     }
     if (theta < (num_samples / 360.0)) { // empirically-determined threshold
       // here we use the Gaussian approximation, but with a modified num_std_devs
-      const unsigned index = 3 * num_samples + (num_std_devs - 1);
+      const unsigned index = 3 * static_cast<unsigned>(num_samples) + (num_std_devs - 1);
       const double raw_ub = cont_classic_ub(num_samples, theta, ub_equiv_table[index]);
       return raw_ub + 0.5; // fake round up
     }
     // This is the most difficult range to approximate; we will compute an "exact" UB.
     // We know that est <= 360, so specialNPrimeF() shouldn't be ridiculously slow.
     const double delta = delta_of_num_std_devs[num_std_devs];
-    return special_n_prime_f(num_samples, theta, delta); // no need to round
+    return static_cast<double>(special_n_prime_f(num_samples, theta, delta)); // no need to round
   }
 
   static void check_theta(double theta) {
